Linting and execution errors
The function "summary" was defined in the r-dml-401k-IV notebook, but then because summary is a global function used by the stats package, redefining it , basically redefined the global variable. Any script that executes after this script, will have the summary function redefined and this throws an error. Better not re-use and re-define such global library R functions.